[cjson] Implement readString
This is not to be confused with string parsing. While it does
handle character range validation and escape sequences, the
parsing class should be agnostic to output encoding. Also, re-
encoding will introduce the difficulty of handling variable
length character encodings which should be left to the data
structure that will eventually hold the strings.